requirements.txt
Django==1.10.7
django-crispy-forms==1.6.1
django-formtools==2.0
django-haystack==2.6.1
future==0.16.0
httplib2==0.10.3
jieba==0.38
Markdown==2.6.8
Pillow==4.2.1
PyMySQL==0.7.11
service-identity==16.0.0
six==1.10.0
Whoosh==2.7.4
django-anymail[sendgrid]==0.10
django-contrib-comments==1.8.0
djangorestframework==3.6.4
social-auth-app-django==1.2.0
django-notifications-hq==1.3
raven==6.2.1
